Sprout FactoryBot::NamespacedFactories package (#830)
In order to share our factories from the staffing center with the job
board without changing every single factory name to have an `sc_`
prefix; we want to make it possible to namespace a factory so they are
conditionally prefixed (i.e. in the staffing center, no prefix. In the
job board; prefix!)

Co-authored-by: Eric Lund <eric@coolworks.com>